pyproject.toml
[project]
name = "sift"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=1.9.0",
    "pillow>=12.2.0",
    "PySide6>=6.9.1",
    "qwen-vl-utils>=0.0.14",
    "torch==2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.57.0",
    "qdrant-client>=1.17.1",
    "easyocr>=1.7.2",
    "faster-whisper>=1.2.1",
    "pandas>=3.0.2",
    "yt-dlp>=2026.3.17",
    "librosa>=0.11.0",
    "soundfile>=0.13.1",
    "watchdog>=6.0.0",
    "pynput>=1.8.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
